ref: Code review refactorings
